!1316 New Features for QPM600 * add default values for dense_mlp_prefetch/lm_head_prefetch from Xu Tianliang * bugfix for not considering A2 machine in deepseek_v3.py * remove redundant code baesd on comments * update 7P1D model config * changes for kv_index_list from Mao Runze * update log level for set_request_id_module in Global Proxy * update 7p1d model configs * changes for lmhead weight nz from Lu Min * changes for top_p/top_k from Gao Yu * changes for omni_attn from Mao Runze * change for omni_attn from Mao Runze * change for deepseek_v3/moe from Xu Tianliang

+++ omni/adaptors/vllm/sample/sampler.py
@@ -242,7 +242,7 @@ class AscendSamplerV1(SamplerV1):
                     q = self.generate_random_sequence(
                         logits, sampling_metadata, spec_metadata,
                     )
-                    res = torch_npu.npu_top_k_top_p_sample(logits, k, p, q)
+                    res = torch_npu.npu_top_k_top_p_sample(logits, k, p, q.to(torch.float32))
                     return res[0]
             else:
                 return apply_top_k_top_p(
